+++ Cargo.toml
@@ -96,6 +96,7 @@ raw-cpuid = "11.5.0"
 testing = []
 mcp = ["dep:rmcp", "rmcp/server", "rmcp/macros", "rmcp/transport-streamable-http-server", "dep:axum", "dep:subtle"]
 cli = ["dep:rmcp", "rmcp/server", "rmcp/macros", "rmcp/client", "rmcp/transport-io", "rmcp/transport-streamable-http-client-reqwest", "dep:clap", "dep:clap_complete"]
+headless = ["cli", "mcp"]
 
 [dev-dependencies]
 egui_kittest = { version = "0.33.3", features = ["eframe"] }
@@ -103,7 +104,7 @@ tokio-shared-rt = "=0.1.0"
 
 [[bin]]
 name = "det-cli"
-path = "src/bin/det_cli.rs"
+path = "src/bin/det_cli/main.rs"
 required-features = ["cli"]
 
 [build-dependencies]
@@ -117,7 +118,7 @@ debug = "line-tables-only"
 
 [lints.rust.unexpected_cfgs]
 level = "warn"
-check-cfg = ["cfg(tokio_unstable)", "cfg(feature, values(\"testing\", \"mcp\", \"cli\"))"]
+check-cfg = ["cfg(tokio_unstable)", "cfg(feature, values(\"testing\", \"mcp\", \"cli\", \"headless\"))"]
 
 [lints.clippy]
 uninlined_format_args = "allow"
